Add routing and Title Page
I also added the html5 routing through $locationProvider. The key to
making that work is to have mirror routes on the web server that render
the main application, in our case '/', '/maxims/', and '/maxims/:id'. The
essence of the idea is that angular handles the routing, but since the
client is going to hit the server first, we have to make sure that the
server spawns the angular app for each available route.