Remove titles from $rootScope;
It's a matter of timing. The titles would have to be appended a
split-second earlier or we would have to wait on some type of change
event which seems like it's more trouble than its worth. The way I have it
now isn't too bad anyway since I keep the titles array in memory.