Only load the maxim ranges that are yet to be loaded.
Before, we were loading both sides of the currently requested id if it
wasn't loaded. But it might be the case that we only need maxims greater
than the current id or maxims less than the current id. In those cases we
are simply wasting bandwidth when we reload the maxims that are already
up.