Add transition animations when switching maxims
This was super tricky. The animation code itself is straight-forward, but
adding the animation classes to dynamic classes was the tricky part.
Basically, you can't rely on using straight up JavaScript to add your
class to the element; if you do, the ng-enter phase won't fire. And you
can't really force a digest or even wait it out with a $timeout. All those
things didn't quite work. The only thing that worked was to apply the
classes via ngClass. When done that way, ng-enter fired correctly. Tricky.