Update prevMaxim and nextMaxim to depend on path
Before, they were simply relying on the value of n which meant that if we
came to that particular maxim via a link, n would be one value and the
current maxim would be another value. So if you clicked the next and prev
arrows, it might jump all the way to what n was before. Now by relying
explicitly on the maxim index in the path string, we can keep the two in
sync.