Fix issue with out of bound maxims not loading
Say for example that we started off on maxim 50. Currently, we would load
all the 10 maxims on either side of the lad, so 40-60. Now, we start
browsing the rest of the maxims and we decide that we want to take a look
at maxim 67. that means we will check the bounds of 50 and 70 and check if
they are loaded. Since 50 is already loaded and 70 is not, we just load up
the maxims from the current maxim to higher bound of 70 (67-70).

Now suppose we then load 65. It won't load. Nor will any maxims in the
range (61-66), because their boundary conditions are met, but we didn't
check if the maxim is actually there. But even then, loading 1 maxim is
generally not worth the hassle. So I find the boundaries dynamically up to
10 starting from index. The tricky case where we wrap around is covered by
the 290-300 and the 1-10 checks.