Remove search bar
There just aren't that many use-cases for searching the book. You would
only really use it if you wanted to find a specific maxim, and you can
look for that using the ToC listing.

The main reason for scrapping it is that the keywords you would search for
aren't apparent. Each maxim uses regular language and it's not unique, so
there is no point in searching it. You just have to look at the sentiment
of the maxim itself through the listings.

I kept the .search styling in the less file, in case I change my mind.
The commented out section is not in the generated css file anyway.