Fix ToC maxim transition on home page
The ToC links relied on there being a "/maxim/n" path string, then we
would match against this and access the first match using an index.
However, on the home page the regex would return null and we would then
try and index null leading to an error and the link doing nothing.

The fix is jus to index the match if there actually is a match and return
null otherwise.