Add footer
Had to make a tradeoff and add the scrollbar permanently. Otherwise
it would crop up during the animations due to overflow, only to disappear
once the new maxim was done loading. This would essentially shake the
page.

So with the addition of the permanent scrollbar, I had to move everything
about 10px over to the left; explicitly for .next-arrow and #footer.